- add new option "make local copy"

+++ resources/lib/launcher.py
@@ -182,6 +182,19 @@ def buildCmd(filenameRows, romCollection, gameRow, escapeCmd, calledFromSkin):
 	for fileNameRow in filenameRows:
 		rom = fileNameRow[0]
 		Logutil.log('rom: ' +str(rom), util.LOG_LEVEL_INFO)
+		
+		if romCollection.makeLocalCopy:
+			localDir = os.path.join(util.getTempDir(), romCollection.name)
+			if os.path.exists(localDir):
+				Logutil.log("Trying to delete local rom files", util.LOG_LEVEL_INFO)    
+				files = os.listdir(localDir)
+				for file in files:
+					os.remove(os.path.join(localDir, file))
+			localRom = os.path.join(localDir, os.path.basename(str(rom)))
+			Logutil.log('Creating local copy: ' + str(localRom), util.LOG_LEVEL_INFO)
+			if xbmcvfs.copy(rom, localRom):
+				Logutil.log('Local copy created', util.LOG_LEVEL_INFO)
+			rom = localRom
 
 		# If it's a .7z file
 		# Don't extract zip files in case of savestate handling and when called From skin
@@ -293,12 +306,17 @@ def handleCompressedFile(filext, rom, romCollection, emuParams):
 	
 	#Note: Trying to delete temporary files (from zip or 7z extraction) from last run
 	#Do this before launching a new game. Otherwise game could be deleted before launch
-	try:
-		Logutil.log("Trying to delete temporary rom files", util.LOG_LEVEL_INFO)	
-		tempDir = util.getTempDir()
-		files = os.listdir(tempDir)
-		for file in files:
-			os.remove(os.path.join(tempDir, file))
+	tempDir = os.path.join(util.getTempDir(), 'extracted')
+	#check if folder exists
+	if(not os.path.isdir(tempDir)):
+		os.mkdir(tempDir)
+	
+	try:		
+		if os.path.exists(tempDir):
+			Logutil.log("Trying to delete temporary rom files", util.LOG_LEVEL_INFO)
+			files = os.listdir(tempDir)
+			for file in files:
+				os.remove(os.path.join(tempDir, file))
 	except Exception, (exc):
 		Logutil.log("Error deleting files after launch emu: " +str(exc), util.LOG_LEVEL_ERROR)
 		gui.writeMsg(util.localize(32036) +": " +str(exc))
@@ -339,7 +357,7 @@ def handleCompressedFile(filext, rom, romCollection, emuParams):
 			Logutil.log('Error handling compressed file', util.LOG_LEVEL_WARNING)
 			return []
 		for archive in archives:					
-			newPath = os.path.join(util.getTempDir(), archive[0])
+			newPath = os.path.join(tempDir, archive[0])
 			fp = open(newPath, 'wb')
 			fp.write(archive[1])
 			fp.close()
@@ -362,14 +380,14 @@ def handleCompressedFile(filext, rom, romCollection, emuParams):
 			Logutil.log('Error handling compressed file', util.LOG_LEVEL_WARNING)
 			return []
 		for archive in archives:
-			newPath = os.path.join(util.getTempDir(), archive[0])
+			newPath = os.path.join(tempDir, archive[0])
 			Logutil.log("Putting extracted file in %s" % newPath, util.LOG_LEVEL_INFO)			
 			fo = open(str(newPath), 'wb')
 			fo.write(archive[1])
 			fo.close()
 		
 		# Point file name to the chosen file and continue as usual
-		roms = [os.path.join(util.getTempDir(), names[chosenROM])]
+		roms = [os.path.join(tempDir, names[chosenROM])]
 		
 	return roms
